pyproject.toml
[project]
name = "marimo-lsp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
authors = [{ name = "Trevor Manz", email = "trevor.j.manz@gmail.com" }]
dependencies = [
  "jupytext>=1.17.2",
  "lsprotocol>=2025.0.0rc1",
  "marimo>=0.15.0",
  "pygls>=2.0.0a4",
  "pyzmq>=27.0.2",
]

[project.scripts]
marimo-lsp = "marimo_lsp:main"

[dependency-groups]
dev = [
  "dirty-equals>=0.9.0",
  "inline-snapshot>=0.27.2",
  "pytest-lsp==1.0.0b2",
  "pytest>=8.4.1",
  "ruff>=0.12.1",
  "ty>=0.0.1a18",
]

[build-system]
requires = ["uv_build>=0.8.7,<0.9.0"]
build-backend = "uv_build"

[tool.ruff]
extend-exclude = ["vscode/*"]

[tool.ruff.lint]
pydocstyle = { convention = "numpy" }
select = ["ALL"]
ignore = [
  "COM812",  # Missing trailing comma (conflicts with ruff format)
  "ISC001",  # Import sorting (conflicts with ruff format)
  "FIX002",  # Fixable issue
  "G004",  # Uses of f-strings in logging
  "PLC2701",  # Private imports
  "TD",  # TODOS
  "S101",  # Use of assert
  "ANN202",  # Return types
]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = [
  "D",  # No docstrings in tests
  "S101",  # Use of assert
  "B018",  # "useless expression", for accessing the Foo._repr_mimbundle_ descriptor
  "SLF001",
  "PLC",
  "PLR",
  "FIX",
]

[tool.uv.sources]
marimo = { path = "../marimo", editable = true }
